refactor(api): change lib directory
