+++ apps/web/src/app/api/play/route.ts
@@ -8,7 +8,7 @@ import {
 	playCounterSchema,
 } from "@wavpoint/utils";
 import { cookies } from "next/headers";
-import { authenticate } from "../../lib/auth";
+import { authenticate } from "../../../lib/auth";
 
 export async function POST(req: Request) {
 	if (
